requirements.txt
yfinance
pandas_gbq
apache-airflow